refactor(contract): split executor/runtime.rs into submodules
Splits the ~5,526-line `contract/executor/runtime.rs` into focused
submodules under `runtime/`, behavior-preserving:

  - runtime/pool.rs          RuntimePool + its ContractExecutor impl
  - runtime/executor_impl.rs the Executor<Runtime> impls

The root `runtime.rs` keeps shared free helpers, types, and the test
modules. Cross-module-called methods are marked `pub(super)`; explicit
`super::` paths in moved code that resolved to `contract::executor`
become `super::super::`.

Source-scrape guards updated for the new layout (assertions unchanged):
  - op_state_manager.rs EMISSION_SITES gains
    `contract/executor/runtime/executor_impl.rs` so the moved
    best-effort-gossip emission stays covered by both scrapes.
  - the `bridged_*` / #4077 inline-related-fetch pin tests are
    repointed to the submodule that now holds the scraped code.
